De-duplicate flood_model/*/*_precipitation.py
